# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de gestion academica",
    "description": "Documentacion generada automaticamente por swagger-autogen",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/estudiante/": {
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "description": "Obtiene la lista de estudiantes con filtros opcionales",
        "parameters": [
          {
            "name": "pais",
            "in": "query",
            "description": "Filtrar por país (insensible a mayúsculas)",
            "type": "string"
          },
          {
            "name": "nombre",
            "in": "query",
            "description": "Filtrar por nombre exacto",
            "type": "string"
          },
          {
            "name": "minEdad",
            "in": "query",
            "description": "Edad mínima requerida",
            "type": "integer"
          },
          {
            "name": "activo",
            "in": "query",
            "description": "Estado del estudiante (true o false)",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "crear un estudiante nuevo",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos para crear un estudiante nuevo",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Juan Perez"
                },
                "pais": {
                  "type": "string",
                  "example": "Costa Rica"
                },
                "edad": {
                  "type": "number",
                  "example": 22
                },
                "notas": {
                  "type": "array",
                  "example": [
                    90,
                    85
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "nombre",
                "pais",
                "edad"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/estudiante/{id}": {
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "description": "Obtiene la informacion de un estudiante en especifico por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante a buscar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "actualizar un estudiante existente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante a actualizar"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Datos a actualizar del estudiante",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Juan Perez"
                },
                "pais": {
                  "type": "string",
                  "example": "Costa Rica"
                },
                "edad": {
                  "type": "number",
                  "example": 22
                },
                "activo": {
                  "type": "boolean",
                  "example": true
                },
                "notas": {
                  "type": "array",
                  "example": [
                    90,
                    85
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "eliminar un estudiante",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante a eliminar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/estudiante/{id}/notas/{notaIndex}": {
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "ver una nota especifica de un estudiante",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del estudiante"
          },
          {
            "name": "notaIndex",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Indice de la nota dentro del arreglo de notas"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/cursos/": {
      "get": {
        "tags": [
          "Cursos"
        ],
        "summary": "ver todos los cursos",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista de cursos",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "array"
                },
                "items": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "object"
                    },
                    "properties": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "number"
                            },
                            "example": {
                              "type": "number",
                              "example": 1
                            }
                          }
                        },
                        "titulo": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "string"
                            },
                            "example": {
                              "type": "string",
                              "example": "programacion"
                            }
                          }
                        },
                        "duracionSemanas": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "number"
                            },
                            "example": {
                              "type": "number",
                              "example": 16
                            }
                          }
                        },
                        "publicado": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "boolean"
                            },
                            "example": {
                              "type": "boolean",
                              "example": true
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cursos"
        ],
        "summary": "crear un producto",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos para crear un curso nuevo",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "Programacion web"
                },
                "duracionSemanas": {
                  "type": "number",
                  "example": 16
                }
              },
              "required": [
                "titulo",
                "duracionSemanas"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/cursos/{id}": {
      "put": {
        "tags": [
          "Cursos"
        ],
        "summary": "actualizar un curso existente",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del curso a actualizar"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Datos a actualizar del curso",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "Programacion web"
                },
                "duracionSemanas": {
                  "type": "number",
                  "example": 16
                },
                "publicado": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Cursos"
        ],
        "summary": "eliminar un curso",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID del curso a eliminar"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}